refactor: regex, string extension
