+++ lib/string_helper.dart
@@ -1,0 +1,7 @@
+extension StringHelper on String {
+  bool isInteger() => int.tryParse(this) != null;
+
+  int toInteger() => int.parse(this);
+
+  bool isNegativeInteger() => isInteger() && startsWith('-');
+}
